Move tests to the tests directory
